+++ IspDataDb/dbo/Tables/ChartBalance.sql
@@ -1,12 +1,14 @@
 ﻿CREATE TABLE [dbo].[ChartBalance] (
     [IdNo]          INT        IDENTITY (1, 1) NOT NULL,
-    [Year]          INT        NOT NULL,
-    [AccountIdNo]   INT        NOT NULL,
+    [Year]          SMALLINT   NOT NULL,
+    [AccountIdNo]   SMALLINT   NOT NULL,
     [Debit]         MONEY      NULL,
     [Credit]        MONEY      NULL,
     [DateTimeStamp] ROWVERSION NULL,
     CONSTRAINT [PK_ChartBalance] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
 
 
 GO
