+++ IspDataDb/dbo/Tables/Designation.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Designation] (
-    [IdNo]               INT            IDENTITY (1, 1) NOT NULL,
+    [IdNo]               SMALLINT       IDENTITY (1, 1) NOT NULL,
     [DesignationCode]    VARCHAR (5)    COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [DesignationName]    VARCHAR (50)   COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [DesignationNameAra] NVARCHAR (100) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -7,4 +7,6 @@
     [DateTimeStamp]      ROWVERSION     NULL,
     CONSTRAINT [PK_Designation] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
