+++ IspDataDb/dbo/Tables/SecurityGroup.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[SecurityGroup] (
-    [IdNo]                 INT           IDENTITY (1, 1) NOT NULL,
+    [IdNo]                 SMALLINT      IDENTITY (1, 1) NOT NULL,
     [SecurityGroupName]    VARCHAR (50)  NULL,
-    [ParentIdNo]           INT           NULL,
+    [ParentIdNo]           SMALLINT      NULL,
     [Notes]                VARCHAR (100) NULL,
     [DateTimeStamp]        ROWVERSION    NULL,
     [SecurityGroupCode]    VARCHAR (10)  NULL,
@@ -9,4 +9,6 @@
     CONSTRAINT [PK_IDNo] PRIMARY KEY CLUSTERED ([IdNo] ASC),
     CONSTRAINT [IX_SecurityGroupName] UNIQUE NONCLUSTERED ([SecurityGroupName] ASC)
 );
+
+
